♻️ refactor : MyReservations api 수정 및 type 수정
api와 type을 수정
api는 최대한 간단하게 구현하고 세세한 에러처리는 페이지에서 하는 걸로 결정

Resolves: #96
